requirements.txt
numpy==1.23.5
pandas==2.0.1
Pillow==9.5.0
Pillow==9.5.0
scikit_learn==1.2.2
streamlit==1.22.0
streamlit_option_menu==0.3.2
